minor fixes float added

+++ app/src/main/java/com/example/customsharedpreference/roomDbSetup/DataTypes.kt
@@ -1,5 +1,5 @@
 package com.example.customsharedpreference.roomDbSetup
 
 enum class DataTypes {
-    STRING, LONG, DOUBLE, INT
+    STRING, LONG, FLOAT, INT
 }
